Remove SplineTest and CatmullRomSpline, rename CatmullRom to CatmullRomSpline.
